setup and configured firbase, fixed outstanding feedbacks
