+++ package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "babel-eslint": "^8.2.2",
     "expo": "^25.0.0",
+    "firebase": "^4.12.0",
     "native-base": "^2.3.10",
     "react": "16.2.0",
     "react-native": "0.52.0",
